refactor: extract hook templates from init.ts to template files
Move hardcoded template strings from init.ts into actual template files:
- templates/hooks/session.ts - session data management
- templates/hooks/index.ts - main hook handlers

This improves maintainability by separating templates from code logic
and makes it easier to modify hook templates without touching the init
command implementation.
