+++ templates/hooks/session.ts
@@ -1,0 +1,34 @@
+import { mkdir } from 'node:fs/promises';
+import { writeFile, readFile } from 'node:fs/promises';
+import * as path from 'path';
+import { tmpdir } from 'node:os';
+
+const SESSIONS_DIR = path.join(tmpdir(), 'claude-hooks-sessions');
+
+export async function saveSessionData(hookType: string, payload: any): Promise<void> {
+  try {
+    // Ensure sessions directory exists
+    await mkdir(SESSIONS_DIR, { recursive: true });
+    
+    const timestamp = new Date().toISOString();
+    const sessionFile = path.join(SESSIONS_DIR, `${payload.session_id}.json`);
+    
+    let sessionData: any[] = [];
+    try {
+      const existing = await readFile(sessionFile, 'utf-8');
+      sessionData = JSON.parse(existing);
+    } catch {
+      // File doesn't exist yet
+    }
+    
+    sessionData.push({
+      timestamp,
+      hookType,
+      payload
+    });
+    
+    await writeFile(sessionFile, JSON.stringify(sessionData, null, 2));
+  } catch (error) {
+    console.error('Failed to save session data:', error);
+  }
+}
